Reformatting and updating `ics06_solomachine` client to resolve bugs and improve performance.
