+++ crates/ibc/src/clients/ics06_solomachine/header.rs
@@ -10,6 +10,7 @@ use prost::Message;
 pub const SOLOMACHINE_HEADER_TYPE_URL: &str = "/ibc.lightclients.solomachine.v1.Header";
 
 /// Header defines a solo machine consensus header
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, PartialEq)]
 pub struct Header {
     /// sequence to update solo machine public key at
