+++ crates/ibc/src/clients/ics06_solomachine/misbehaviour/signature_and_data.rs
@@ -1,11 +1,12 @@
-use crate::clients::ics06_solomachine::data_type::DataType;
 use crate::clients::ics06_solomachine::error::Error;
+use crate::clients::ics06_solomachine::types::DataType;
 use crate::prelude::*;
 use ibc_proto::ibc::lightclients::solomachine::v1::SignatureAndData as RawSignatureAndData;
 use ibc_proto::protobuf::Protobuf;
 
 /// SignatureAndData contains a signature and the data signed over to create that
 /// signature.
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, PartialEq)]
 pub struct SignatureAndData {
     pub signature: Vec<u8>,
