+++ crates/ibc/src/clients/ics06_solomachine/mod.rs
@@ -1,6 +1,6 @@
 pub mod client_state;
 pub mod consensus_state;
-pub mod data_type;
 pub mod error;
 pub mod header;
 pub mod misbehaviour;
+pub mod types;
